feat: enable new VAI page
