--- apps/evm/src/pages/Vai/MintVai/index.tsx
@@ -1,16 +1,0 @@
-import { ConnectWallet } from 'containers/ConnectWallet';
-import { useTranslation } from 'libs/translations';
-
-import { Form } from './Form';
-
-const MintVai: React.FC = () => {
-  const { t } = useTranslation();
-
-  return (
-    <ConnectWallet message={t('vai.mintVai.connectWallet')}>
-      <Form />
-    </ConnectWallet>
-  );
-};
-
-export default MintVai;
