--- apps/evm/src/pages/Vai/constants.ts
@@ -1,1 +1,0 @@
-export const VAI_ID = 'vai';
