--- apps/evm/src/pages/Vai/testIds.ts
@@ -1,7 +1,0 @@
-const TEST_IDS = {
-  repayTextField: 'vai-repay-text-field',
-  spendingLimit: 'vai-spending-limit',
-  primeOnlyWarning: 'vai-prime-only-warning',
-};
-
-export default TEST_IDS;
